FIXED Movement/attacking, Implemented score saving under a name, added sound effects

+++ Space Strategy/Space Strategy/Ship.h
@@ -169,7 +169,7 @@ protected:
 	ShipType m_shipType;
 	unsigned int m_id;
 	bool m_hasUpdatedOnce = false;
-	int m_moveCounter;
+	float m_moveCounter;
 	bool m_moveFinished = false;
 
 	Fleet* m_fleet;
